整合进example与model，添加未补充结果的readme
